Refactor and fix some testing helpers
